// class-room-sauvegarde/12/src/app/api/model/classeEleveDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SousClasseDto } from './sousClasseDto';
import { ClasseDto } from './classeDto';
import { EleveDto } from './eleveDto';


export interface ClasseEleveDto { 
    idClasseEleve?: number;
    eleve?: EleveDto;
    classe?: ClasseDto;
    sousClasse?: SousClasseDto| null; // Allow null;
}